refactor(user_avatar_color): move user avatar color computation to a separate file
